+++ CommNetAntennasInfo/Properties/AssemblyVersion.cs
@@ -4,4 +4,4 @@
   
  using System.Reflection;
 
- [assembly: AssemblyVersion("2.2.0.0")]
+ [assembly: AssemblyVersion("2.3.0.0")]